% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrrOG.R
\name{GrrgOG}
\alias{GrrgOG}
\title{The Greg OG Soundboard Package}
\description{
A collection of Dr. Greg Matthews PhDizzle's iconic noises, reactions,
and occasional statistics references.
This package allows users to play various sound clips with simple function calls.

Key functions in this package include:
\itemize{
  \item \code{\link{play_OG}}: Plays a sound from the "OG phrases" collection.
  \item \code{\link{play_stats}}: Plays a sound from the "Some statistics" collection.
  \item \code{\link{play_Gmisc}}: Plays a sound from the "Reactionary noises" collection.
}

Code is modeled after \code{skrrrahh} from the BRRR package (\link{https://github.com/brooke-watson/BRRR/blob/master/R/skrrrahh.R}).
}
\examples{
# Play the first sound in the OG phrases collection
play_OG(1)

# Play the second sound in the stats collection
play_stats(2)

# Play a random reactionary noise
play_Gmisc(10)
}
