% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrrOG.R
\name{OG_phrases}
\alias{OG_phrases}
\title{The Greg OG Soundboard Package}
\usage{
OG_phrases()
}
\arguments{
\item{index}{Numeric. The index of the audio file to play.}
}
\value{
Plays an audio file but returns NULL.
}
\description{
A collection of Dr. Greg Matthews PhDizzle's iconic noises, reactions,
and occasional statistics references.

This package allows users to play various sound clips with simple function calls.

Code is modeled after \code{skrrrahh} from the BRRR package (https://github.com/brooke-watson/BRRR/blob/master/R/skrrrahh.R).
}
\examples{
play_OG(1)  # Plays the first audio file from the OG_phrases audio list.
play_stats(5)  # Plays the fifth audio file from the some_stats audio list.
play_Gmisc(10) # Plays the tenth audio file from the Gmisc audio list.

}
