% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrrOG.R
\name{play_Gmisc}
\alias{play_Gmisc}
\title{Play a Miscellaneous Sound Clip}
\usage{
play_Gmisc(index)
}
\arguments{
\item{index}{Numeric. The index of the audio file to play.}
}
\value{
NULL (plays an audio file).
}
\description{
This function plays a sound clip from the "Gmisc" list.
}
\examples{
play_Gmisc(1)  # Plays the first miscellaneous phrase.

}
