% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrrOG.R
\name{play_OG}
\alias{play_OG}
\title{Play an OG Sound Clip}
\usage{
play_OG(index)
}
\arguments{
\item{index}{Numeric. The index of the audio file to play.}
}
\value{
NULL (plays an audio file).
}
\description{
This function plays a sound clip from the "OG phrases" list.
}
\examples{
play_OG(1)  # Plays the first OG phrase.

}
