% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrrOG.R
\name{play_stats}
\alias{play_stats}
\title{Play a Statistics-Related Sound Clip}
\usage{
play_stats(index)
}
\arguments{
\item{index}{Numeric. The index of the audio file to play.}
}
\value{
NULL (plays an audio file).
}
\description{
This function plays a sound clip from the "some statistics" list.
}
\examples{
play_stats(1)  # Plays the first statistics-related phrase.
}
