% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sq_function.R
\name{squares}
\alias{squares}
\title{Square of x
This function generate the square of the given value.}
\usage{
squares(x)
}
\arguments{
\item{x}{any number}
}
\description{
Square of x
This function generate the square of the given value.
}
\examples{
squares()
}
\keyword{squares}
