% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrrOG.R
\name{GrrgOG}
\alias{GrrgOG}
\title{The Greg OG Soundboard Package}
\description{
A collection of Dr. Greg Matthews PhDizzle's iconic noises, reactions,
and occasional statistics references.

This package allows users to play various sound clips with simple function calls.

Code is modeled after \code{skrrrahh} from the BRRR package (https://github.com/brooke-watson/BRRR/blob/master/R/skrrrahh.R).
}
