% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GrrOG.R
\name{play_OG}
\alias{play_OG}
\title{The Greg OG Soundboard Package}
\usage{
play_OG(index)
}
\arguments{
\item{index}{Numeric. The index of the audio file to play.}
}
\value{
NULL (plays an audio file).
}
\description{
A collection of Dr. Greg Matthews PhDizzle's iconic noises, reactions,
and occasional statistics references.

This package allows users to play various sound clips with simple function calls.

Code is modeled after \code{skrrrahh} from the BRRR package (https://github.com/brooke-watson/BRRR/blob/master/R/skrrrahh.R).

Play an OG Sound Clip

This function plays a sound clip from the "OG phrases" list.
}
\examples{
play_OG(1)  # Plays the first OG phrase.
}
